requirements.txt
Flask==3.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy-Utils==0.41.2